refactor: Move ResourceBundle helper classes to util.resources.
This is to make it simple to verify the transit model dependencies, the o.o.util
contains a lot of none util classes, so a dependency to that is not allowed.
