+++ src/main/java/org/opentripplanner/api/resource/BikeRental.java
@@ -19,7 +19,7 @@ import org.opentripplanner.routing.vehicle_rental.VehicleRentalPlace;
 import org.opentripplanner.routing.vehicle_rental.VehicleRentalStationService;
 import org.opentripplanner.standalone.server.OTPServer;
 import org.opentripplanner.standalone.server.Router;
-import org.opentripplanner.util.ResourceBundleSingleton;
+import org.opentripplanner.util.resources.ResourceBundleSingleton;
 
 @Path("/routers/{ignoreRouterId}/bike_rental")
 public class BikeRental {
